Position block mover below the block for floats, wide blocks.
This PR is inspired by recent discussions around how to handle the block mover in wide context for themes that heavily style the editor, like TwentyNineteen. As such it aims to find a more scalable way of showing block side-UI for blocks that are wider than the main content, or for blocks that have special requirements such as floats.

What this PR does is a few things:

- It moves the side UI to be _below_ the block, for floats, wide and full-wide blocks.
- It changes the behavior so this UI does not show on hover, but it is permanently visible when the block is selected.
- It makes it so when one of these blocks are selected, adjacent (subsequent) blocks are pushed downwards, to make room for the controls.

In doing so it also mitigates the oft-reported stress-case, which is that the side UI is currently not visible at all for floated blocks. Most importantly, however, it should allow more creative layouts for themes that heavily style the editor, since the block mover is now consistently in only two places — on the side for normal blocks, and below the block for any alignment other than centered.