Add timezone to PostSchedule date
Fixes #15673

Removes the settings format and replaces with a date format that
includes the timezone. The setting inside the date format package is
equally hard-coded and not from wp-admin settings configuration.

It makes sense to me to hard-code the setting closer to the usage, I'm
open to alternatives, issue #15673 highlights a couple of different
solutions considered but did not receive any feedback.